metachat-main/web-app/pyproject.toml
[project]
name = "metachat-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Robert Lupoiu"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.60.1,<2.0.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "scipy (>=1.15.1,<2.0.0)",
    "sympy (>=1.13.3,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "gdspy (>=1.6.13,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
